requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
fastapi==0.116.1
filelock==3.19.1
fsspec==2025.9.0
gTTS==2.5.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
kombu==5.5.4
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
openai-whisper==20250625
packaging==25.0
prompt_toolkit==3.0.52
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
regex==2025.9.1
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
sympy==1.14.0
tiktoken==0.11.0
torch==2.8.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
yt-dlp==2025.8.27
